move util functions into util.py
